requirements.txt
opencv-python
mediapipe
tensorflow
pyttsx3
pillow